Use a multi-stage approach to squash all the layers
This seems like a good compromise between keeping layers in the cache
for easier development, and reducing final image size.  This results
in a 1.05GB image for me, down from 1.3GB today (which is being lax
about layers).  I tried moving basically all the build steps in the
Dockerfile to a script, ending up with a single RUN invocation, and
that reduces the image further to 940MB, but, with that approach,
every time you experiment with installing an extra package, docker
must rebuild the whole thing.